fixed majority of tests so can isolate test stuff from FastDifferentiation package
